[REV] update delete backup method

+++ console/services/backup_service.py
@@ -123,9 +123,11 @@ class GroupAppBackupService(object):
             return 404, "不存在该备份记录"
         if backup_record.status == "starting":
             return 409, "该备份正在进行中"
+        if backup_record.status == "success":
+            return 409, "该备份不可删除"
         region_api.delete_backup_by_backup_id(region, tenant.tenant_name, backup_id)
         backup_record_repo.delete_record_by_backup_id(tenant.tenant_id, backup_id)
-        return 200,"success"
+        return 200, "success"
 
     def get_group_backup_status_by_group_id(self, tenant, region, group_id):
         backup_records = backup_record_repo.get_record_by_group_id(group_id)
